+++ Service.cs
@@ -3,12 +3,14 @@ using Dalamud.Game;
 using Dalamud.IoC;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
+#pragma warning disable CA1416 // Validate platform compatibility
 
 namespace ActionTimelineReplacement;
 
 internal class Service
 {
     internal static Configuration Config { get; set; } = null!;
+
 
     [PluginService] internal static IDalamudPluginInterface PluginInterface { get; set; } = null!;
     [PluginService] internal static ISigScanner Scanner { get; set; } = null!;
