+++ Sheets/Action/ActionManager.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Runtime.InteropServices;
 using Lumina.Excel.Sheets;
+#pragma warning disable CA1416 // Validate platform compatibility
 
 namespace ActionTimelineReplacement.Sheets;
