+++ Sheets/Glasses/GlassesMain.cs
@@ -1,32 +1,24 @@
 ﻿using Dalamud.Interface.Utility.Raii;
 using Dalamud.Interface;
-using System.Numerics;
 using ImGuiNET;
 using System;
 using System.Linq;
-using Dalamud.Interface.Utility;
 using ActionTimelineReplacement.Base.Setups;
 using ActionTimelineReplacement.Windows;
+using ActionTimelineReplacement.Base.Items.Global;
 
 namespace ActionTimelineReplacement.Sheets;
 
 #region Main
 public class GlassesMain
 {
-    private static float Scale => ImGuiHelpers.GlobalScale;
     public static void Draw(string mainkey, ref Configuration.ReplacementSet _activeSet, ref string search)
     {
-        var itemHeight = ImGui.CalcTextSize("").Y + ImGui.GetStyle().FramePadding.Y * 2 + ImGui.GetStyle().WindowPadding.Y;
-
-        //to fix: scale height according to item count
-        using var subList = ImRaii.Child(mainkey, new Vector2(-1, ImGui.GetWindowSize().Y - ImGui.GetCursorPosY() - itemHeight - ImGui.GetStyle().WindowPadding.Y), false);
+        using var subList = ImRaii.Child(mainkey, CalcGlobals.BodyScale(), false);
         if (subList)
         {
             const string searchPopup = "Search glasses";
-            if (ImGui.Button(" + "))
-            {
-                ImGui.OpenPopup(searchPopup);
-            }
+            UiGlobals.DrawAddItem(searchPopup);
 
             foreach (var key in _activeSet.GlassesWriter.Keys)
             {
@@ -49,46 +41,36 @@ public class GlassesMain
                 ImGui.SameLine();
                 ImGui.TextWrapped(GlassesManager.GetName(key));
 
-                //to do: streamline this
-                ImGui.TextUnformatted("Unknown70_1");
-                DrawSByte("Unknown70_1", ref replace.Unknown70_1, i => i.Unknown70_1);
+                DrawSByte("Unknown70_1", "Unknown70_1", ref replace.Unknown70_1, i => i.Unknown70_1);
 
-                ImGui.TextUnformatted("Unknown70_2");
-                DrawSByte("Unknown70_2", ref replace.Unknown70_2, i => i.Unknown70_2);
+                DrawSByte("Unknown70_2", "Unknown70_2", ref replace.Unknown70_2, i => i.Unknown70_2);
 
-                ImGui.TextUnformatted("Unknown70_3");
-                DrawSByte("Unknown70_3", ref replace.Unknown70_3, i => i.Unknown70_3);
+                DrawSByte("Unknown70_3", "Unknown70_3", ref replace.Unknown70_3, i => i.Unknown70_3);
 
-                ImGui.TextUnformatted("Unknown70_4");
-                DrawSByte("Unknown70_4", ref replace.Unknown70_4, i => i.Unknown70_4);
+                DrawSByte("Unknown70_4", "Unknown70_4", ref replace.Unknown70_4, i => i.Unknown70_4);
 
-                ImGui.TextUnformatted("Unknown70_5");
-                DrawSByte("Unknown70_5", ref replace.Unknown70_5, i => i.Unknown70_5);
+                DrawSByte("Unknown70_5", "Unknown70_5", ref replace.Unknown70_5, i => i.Unknown70_5);
 
-                ImGui.TextUnformatted("Unknown70_6");
-                DrawSByte("Unknown70_6", ref replace.Unknown70_6, i => i.Unknown70_6);
+                DrawSByte("Unknown70_6", "Unknown70_6",ref replace.Unknown70_6, i => i.Unknown70_6);
 
-                ImGui.TextUnformatted("Unknown70_7");
-                DrawUInt("Unknown70_7", ref replace.Unknown70_7, i => i.Unknown70_7);
+                DrawUInt("Unknown70_7", "Unknown70_7", ref replace.Unknown70_7, i => i.Unknown70_7);
 
-                ImGui.TextUnformatted("Unknown70_8");
-                DrawUShort("Unknown70_8", ref replace.Unknown70_8, i => i.Unknown70_8);
+                DrawUShort("Unknown70_8", "Unknown70_8", ref replace.Unknown70_8, i => i.Unknown70_8);
 
-                ImGui.NewLine();
-                ImGui.Separator();
-                ImGui.NewLine();
+                UiGlobals.DrawItemSeparator();
                 continue;
 
                 #endregion
                 #region Items
                 
-                void DrawUShort(string name, ref ushort value,
+                void DrawUShort(string refname, string text, ref ushort value,
                     Func<GlassesReplace, ushort> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (ushort)relay;
                         Setup.SetGlasses(key);
@@ -98,7 +80,7 @@ public class GlassesMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(GlassesManager.GetOriginal(key));
                             Setup.SetGlasses(key);
@@ -107,13 +89,13 @@ public class GlassesMain
                     }
                 }
 
-                void DrawSByte(string name, ref sbyte value,
+                void DrawSByte(string refname, string text, ref sbyte value,
                      Func<GlassesReplace, sbyte> getDefault)
                 {
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (sbyte)relay;
                         Setup.SetGlasses(key);
@@ -123,7 +105,7 @@ public class GlassesMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(GlassesManager.GetOriginal(key));
                             Setup.SetGlasses(key);
@@ -132,13 +114,13 @@ public class GlassesMain
                     }
                 }
 
-                void DrawUInt(string name, ref uint value,
+                void DrawUInt(string refname, string text, ref uint value,
                     Func<GlassesReplace, uint> getDefault)
                 {
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = (int)value;
-                    if (ImGui.DragInt("##" + name + key, ref relay))
+                    if (ImGui.DragInt("##" + refname + key, ref relay))
                     {
                         value = (uint)relay;
                         Setup.SetGlasses(key);
@@ -148,7 +130,7 @@ public class GlassesMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(GlassesManager.GetOriginal(key));
                             Setup.SetGlasses(key);
@@ -164,14 +146,11 @@ public class GlassesMain
             using var searchGlasses = ImRaii.Popup(searchPopup);
             if (searchGlasses)
             {
-                var width = 200 * Scale;
-                var height = 200 * Scale;
-
-                ImGui.SetNextItemWidth(width);
+                ImGui.SetNextItemWidth(CalcGlobals.XY());
                 ImGui.InputText("##Search glasses", ref search, 256);
                 var localsearch = search;
 
-                using var popupChild = ImRaii.Child(searchPopup, new Vector2(width, height), true);
+                using var popupChild = ImRaii.Child(searchPopup, CalcGlobals.SearchPopScale(), true);
                 foreach (var pair in GlassesManager.Names.OrderBy(i =>
                 {
                     if (string.IsNullOrEmpty(localsearch)) return 0;
