+++ Sheets/Status/StatusMain.cs
@@ -1,34 +1,25 @@
 ﻿using Dalamud.Interface.Utility.Raii;
 using Dalamud.Interface;
-using System.Numerics;
 using ImGuiNET;
 using System;
 using System.Linq;
-using Dalamud.Interface.Utility;
 using ActionTimelineReplacement.Base.Setups;
 using ActionTimelineReplacement.Windows;
-using Lumina.Excel.Sheets;
+using ActionTimelineReplacement.Base.Items.Global;
+#pragma warning disable CA1416 // Validate platform compatibility
 
 namespace ActionTimelineReplacement.Sheets;
 
 #region Main
 public class StatusMain
 {
-    private static float Scale => ImGuiHelpers.GlobalScale;
     public static void Draw(string mainkey, ref Configuration.ReplacementSet _activeSet, ref string search)
     {
-        var itemHeight = ImGui.CalcTextSize("").Y + ImGui.GetStyle().FramePadding.Y * 2 + ImGui.GetStyle().WindowPadding.Y;
-
-        //to fix: scale height according to item count
-        using var subList = ImRaii.Child(mainkey, new Vector2(-1, ImGui.GetWindowSize().Y - ImGui.GetCursorPosY() - itemHeight - ImGui.GetStyle().WindowPadding.Y), false);
-
+        using var subList = ImRaii.Child(mainkey, CalcGlobals.BodyScale(), false);
         if (subList)
         {
             const string searchPopup = "Search statuses";
-            if (ImGui.Button(" + "))
-            {
-                ImGui.OpenPopup(searchPopup);
-            }
+            UiGlobals.DrawAddItem(searchPopup);
 
             foreach (var key in _activeSet.StatusWriter.Keys)
             {
@@ -52,56 +43,43 @@ public class StatusMain
                 ImGui.SameLine();
                 ImGui.TextWrapped(StatusManager.GetName(key));
 
-                //to do: streamline this
                 //to do: show loop vfx and hit effect as strings
-                ImGui.TextUnformatted("Parameter Modifier");
-                DrawInt("ParamModifier", ref replace.ParamModifier, i => i.ParamModifier);
+                DrawInt("ParamModifier", "Parameter Modifier", ref replace.ParamModifier, i => i.ParamModifier);
 
-                ImGui.TextUnformatted("Status Loop VFX ID");
-                DrawUShort("StatusLoopVFX", ref replace.StatusLoopVFX, i => i.StatusLoopVFX);
+                DrawUShort("StatusLoopVFX", "Status Loop VFX ID", ref replace.StatusLoopVFX, i => i.StatusLoopVFX);
 
-                ImGui.TextUnformatted("Unknown 0");
-                DrawByte("Unknown0", ref replace.Unknown0, i => i.Unknown0);
+                DrawByte("Unknown0", "Unknown 0", ref replace.Unknown0, i => i.Unknown0);
 
-                ImGui.TextUnformatted("Status Category");
-                DrawByte("StatusCategory", ref replace.StatusCategory, i => i.StatusCategory);
+                DrawByte("StatusCategory", "Status Category", ref replace.StatusCategory, i => i.StatusCategory);
 
-                ImGui.TextUnformatted("Status Hit Effect ID");
-                DrawByte("StatusHitEffect", ref replace.StatusHitEffect, i => i.StatusHitEffect);
+                DrawByte("StatusHitEffect", "Status Hit Effect ID", ref replace.StatusHitEffect, i => i.StatusHitEffect);
 
-                ImGui.TextUnformatted("Parameter Effect");
-                DrawByte("ParamEffect", ref replace.ParamEffect, i => i.ParamEffect);
+                DrawByte("ParamEffect", "Parameter Effect", ref replace.ParamEffect, i => i.ParamEffect);
 
-                ImGui.TextUnformatted("Target Type");
-                DrawByte("TargetType", ref replace.TargetType, i => i.TargetType);
+                DrawByte("TargetType", "Target Type", ref replace.TargetType, i => i.TargetType);
 
-                ImGui.TextUnformatted("Flag 1");
-                DrawByte("Flags", ref replace.Flags, i => i.Flags);
+                DrawByte("Flags", "Flag 1", ref replace.Flags, i => i.Flags);
 
-                ImGui.TextUnformatted("Flag 2");
-                DrawByte("Flag2", ref replace.Flag2, i => i.Flag2);
+                DrawByte("Flag2", "Flag 2", ref replace.Flag2, i => i.Flag2);
 
-                ImGui.TextUnformatted("Unknown 70_1");
-                DrawByte("Unknown_70_1", ref replace.Unknown_70_1, i => i.Unknown_70_1);
+                DrawByte("Unknown_70_1", "Unknown_70_1", ref replace.Unknown_70_1, i => i.Unknown_70_1);
 
-                ImGui.TextUnformatted("(related to attack type)");
-                DrawSByte("AtkType", ref replace.Unknown2, i => i.Unknown2);
+                DrawSByte("AtkType", "(related to attack type)", ref replace.Unknown2, i => i.Unknown2);
 
-                ImGui.NewLine();
-                ImGui.Separator();
-                ImGui.NewLine();
+                UiGlobals.DrawItemSeparator();
                 continue;
 
                 #endregion
                 #region Items
 
-                void DrawInt(string name, ref int value,
+                void DrawInt(string refname, string text, ref int value,
                     Func<StatusReplace, int> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = relay;
                         Setup.SetStatus(key);
@@ -111,7 +89,7 @@ public class StatusMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(StatusManager.GetOriginal(key));
                             Setup.SetStatus(key);
@@ -120,13 +98,14 @@ public class StatusMain
                     }
                 }
 
-                void DrawUShort(string name, ref ushort value,
+                void DrawUShort(string refname, string text, ref ushort value,
                     Func<StatusReplace, ushort> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (ushort)relay;
                         Setup.SetStatus(key);
@@ -136,7 +115,7 @@ public class StatusMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(StatusManager.GetOriginal(key));
                             Setup.SetStatus(key);
@@ -145,13 +124,14 @@ public class StatusMain
                     }
                 }
 
-                void DrawByte(string name, ref byte value,
-                     Func<StatusReplace, byte> getDefault)
+                void DrawByte(string refname, string text, ref byte value,
+                    Func<StatusReplace, byte> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (byte)relay;
                         Setup.SetStatus(key);
@@ -161,7 +141,7 @@ public class StatusMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(StatusManager.GetOriginal(key));
                             Setup.SetStatus(key);
@@ -170,13 +150,14 @@ public class StatusMain
                     }
                 }
 
-                void DrawSByte(string name, ref sbyte value,
-                     Func<StatusReplace, sbyte> getDefault)
+                void DrawSByte(string refname, string text, ref sbyte value,
+                    Func<StatusReplace, sbyte> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (sbyte)relay;
                         Setup.SetStatus(key);
@@ -186,7 +167,7 @@ public class StatusMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(StatusManager.GetOriginal(key));
                             Setup.SetStatus(key);
@@ -202,14 +183,11 @@ public class StatusMain
             using var searchStatus = ImRaii.Popup(searchPopup);
             if (searchStatus)
             {
-                var width = 200 * Scale;
-                var height = 200 * Scale;
-
-                ImGui.SetNextItemWidth(width);
+                ImGui.SetNextItemWidth(CalcGlobals.XY());
                 ImGui.InputText("##Search statuses", ref search, 256);
                 var localsearch = search;
 
-                using var popupChild = ImRaii.Child(searchPopup, new Vector2(width, height), true);
+                using var popupChild = ImRaii.Child(searchPopup, CalcGlobals.SearchPopScale(), true);
                 foreach (var pair in StatusManager.Names.OrderBy(i =>
                 {
                     if (string.IsNullOrEmpty(localsearch)) return 0;
