+++ Sheets/TiltParam/TiltParamMain.cs
@@ -1,33 +1,25 @@
 ﻿using Dalamud.Interface.Utility.Raii;
 using Dalamud.Interface;
-using System.Numerics;
 using ImGuiNET;
 using System;
 using System.Linq;
-using Dalamud.Interface.Utility;
 using ActionTimelineReplacement.Base.Setups;
 using ActionTimelineReplacement.Windows;
+using ActionTimelineReplacement.Base.Items.Global;
+#pragma warning disable CA1416 // Validate platform compatibility
 
 namespace ActionTimelineReplacement.Sheets;
 
 #region Main
 public class TiltParamMain
 {
-    private static float Scale => ImGuiHelpers.GlobalScale;
     public static void Draw(string mainkey, ref Configuration.ReplacementSet _activeSet, ref string search)
     {
-        var itemHeight = ImGui.CalcTextSize("").Y + ImGui.GetStyle().FramePadding.Y * 2 + ImGui.GetStyle().WindowPadding.Y;
-
-        //to fix: scale height according to item count
-        using var subList = ImRaii.Child(mainkey, new Vector2(-1, ImGui.GetWindowSize().Y - ImGui.GetCursorPosY() - itemHeight - ImGui.GetStyle().WindowPadding.Y), false);
-
+        using var subList = ImRaii.Child(mainkey, CalcGlobals.BodyScale(), false);
         if (subList)
         {
             const string searchPopup = "Search tilts";
-            if (ImGui.Button(" + "))
-            {
-                ImGui.OpenPopup(searchPopup);
-            }
+            UiGlobals.DrawAddItem(searchPopup);
 
             foreach (var key in _activeSet.TiltParamWriter.Keys)
             {
@@ -51,40 +43,32 @@ public class TiltParamMain
                 ImGui.SameLine();
                 ImGui.TextWrapped(TiltParamManager.GetName(key));
 
-                //to do: streamline this
-                ImGui.TextUnformatted("Tilt Rate");
-                DrawUShort("TiltRate", ref replace.Unknown0, i => i.Unknown0);
+                DrawUShort("TiltRate", "Tilt Rate", ref replace.Unknown0, i => i.Unknown0);
 
-                ImGui.TextUnformatted("Rotation Origin Offset");
-                DrawByte("RotOriginOffset", ref replace.Unknown1, i => i.Unknown1);
+                DrawByte("RotOriginOffset", "Rotation Origin Offset", ref replace.Unknown1, i => i.Unknown1);
 
-                ImGui.TextUnformatted("Max Angle");
-                DrawByte("MaxAngle", ref replace.Unknown2, i => i.Unknown2);
+                DrawByte("MaxAngle", "Max Angle", ref replace.Unknown2, i => i.Unknown2);
 
-                ImGui.TextUnformatted("Unknown 3");
-                DrawByte("Unknown 3", ref replace.Unknown3, i => i.Unknown3);
+                DrawByte("Unknown3", "Unknown 3", ref replace.Unknown3, i => i.Unknown3);
 
-                ImGui.TextUnformatted("Unknown 4");
-                DrawByte("Unknown 4", ref replace.Unknown4, i => i.Unknown4);
+                DrawByte("Unknown4", "Unknown 4", ref replace.Unknown4, i => i.Unknown4);
 
-                ImGui.TextUnformatted("Reverse Mouse Dir");
-                DrawBool("MouseReverse", ref replace.Unknown5, i => i.Unknown5);
+                DrawBool("MouseReverse", "Reverse Mouse Direction", ref replace.Unknown5, i => i.Unknown5);
 
-                ImGui.NewLine();
-                ImGui.Separator();
-                ImGui.NewLine();
+                UiGlobals.DrawItemSeparator();
                 continue;
 
                 #endregion
                 #region Items
 
-                void DrawUShort(string name, ref ushort value,
+                void DrawUShort(string refname, string text, ref ushort value,
                     Func<TiltParamReplace, ushort> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (ushort)relay;
                         Setup.SetTiltParam(key);
@@ -94,7 +78,7 @@ public class TiltParamMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(TiltParamManager.GetOriginal(key));
                             Setup.SetTiltParam(key);
@@ -103,13 +87,14 @@ public class TiltParamMain
                     }
                 }
 
-                void DrawByte(string name, ref byte value,
+                void DrawByte(string refname, string text, ref byte value,
                     Func<TiltParamReplace, byte> getDefault)
                 {
+                    ImGui.TextUnformatted(text);
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     int relay = value;
-                    if (ImGui.InputInt("##" + name + key, ref relay))
+                    if (ImGui.InputInt("##" + refname + key, ref relay))
                     {
                         value = (byte)relay;
                         Setup.SetTiltParam(key);
@@ -119,7 +104,7 @@ public class TiltParamMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(TiltParamManager.GetOriginal(key));
                             Setup.SetTiltParam(key);
@@ -128,13 +113,13 @@ public class TiltParamMain
                     }
                 }
 
-                void DrawBool(string name, ref bool value,
+                void DrawBool(string refname, string text, ref bool value,
                     Func<TiltParamReplace, bool> getDefault)
                 {
                     ImGui.SetCursorPosX(ImGui.GetCursorPosX() + 15);
-                    ImGui.SetNextItemWidth(110 * Scale);
+                    ImGui.SetNextItemWidth(110 * CalcGlobals.GlobalScale());
                     bool relay = value;
-                    if (ImGui.Checkbox("##" + name + key, ref relay))
+                    if (ImGui.Checkbox("##" + refname + key, ref relay))
                     {
                         value = relay;
                         Setup.SetTiltParam(key);
@@ -144,7 +129,7 @@ public class TiltParamMain
 
                     using (ImRaii.PushFont(UiBuilder.IconFont))
                     {
-                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{name}{key}"))
+                        if (ImGui.Button($"{FontAwesomeIcon.Reply.ToIconString()}##{refname}{key}"))
                         {
                             value = getDefault(TiltParamManager.GetOriginal(key));
                             Setup.SetTiltParam(key);
@@ -160,14 +145,11 @@ public class TiltParamMain
             using var searchTiltParam = ImRaii.Popup(searchPopup);
             if (searchTiltParam)
             {
-                var width = 200 * Scale;
-                var height = 200 * Scale;
-
-                ImGui.SetNextItemWidth(width);
+                ImGui.SetNextItemWidth(CalcGlobals.XY());
                 ImGui.InputText("##Search tilt parameters", ref search, 256);
                 var localsearch = search;
 
-                using var popupChild = ImRaii.Child(searchPopup, new Vector2(width, height), true);
+                using var popupChild = ImRaii.Child(searchPopup, CalcGlobals.SearchPopScale(), true);
                 foreach (var pair in TiltParamManager.Names.OrderBy(i =>
                 {
                     if (string.IsNullOrEmpty(localsearch)) return 0;
